Threading stability changes
Standardise access to AddinHost and sub-objects to try and ensure
correct threading
Async was a pool of threadpools each of size one, combine some into a
pool which is allowed to grow larger. Async also allowed arbitrary names
for the pools that was allowing effective thread leaks when item names
where used as part of the pool name.
Accessing UI from non-UI threads was causing stability issues on xbox.
This was especially noticable when media was streaming which now appears
stable.