Remove excessive image loading
Setting AsyncImageLoader.Size doesn't do anything as the Library image
has previously been modified to no longer perform resizing or caching of
images at specific sizes. This means that PrimaryImage and
SmallPrimaryImage are the same and changing PreferredImageSmallSize has
no actual affect on the image that is loaded only on the
SmallImageDistorted property. This change ensures we don't load/reload
the image more than needed as it was causing a performance hit.