Re-enable caching of resized images
Re-enable caching of resized images, this used to be present in MB2 but
had been removed. The result was that extenders were very slow to load
certain views especially on wifi connections as images are sent to the
extenders raw and resized there to be displayed. Resizing to the right
size on the media center host dramatically improves the extender
experience. In my example a folder that used to take 20secs now takes
3secs to open.