Added an option to enable display of complete series data when a series is selected from the recent list
When selecting a series in the TV section of the EHS (recent list), a
view is presented which contains the recent items only.

Depending on personal preference, some people (like me) would prefer to
see a view covering the complete series. Especially, when the "Custom
Series View" of the Chocolate theme is used.
I added an option to achieve this behaviour. The default is off.

Please see also:
http://emby.media/community/index.php?/topic/25640-chocolate-enhanced-series-view/